pyproject.toml
[project]
name = "holespawn"
version = "0.1.0"
description = "Rabbit hole / ARG generator from Twitter/X (and optional Discord)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "vaderSentiment>=3.3.2",
    "spacy>=3.7.0",
    "nltk>=3.8.0",
    "scikit-learn>=1.3.0",
    "gensim>=4.3.0",
    "anthropic>=0.39.0",
    "openai>=1.0.0",
    "google-genai>=0.5.0",
    "playwright>=1.40",
    "discord.py>=2.0.0",
    "tweepy>=4.0.0",
    "loguru>=0.7.0",
    "tqdm>=4.66.0",
    "pyyaml>=6.0",
    "jsonschema>=4.0.0",
    "networkx>=3.0",
    "python-louvain>=0.16",
    "numpy>=1.20.0",
    "scipy>=1.7.0",
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
    "sqlalchemy>=2.0.0",
    "bcrypt>=4.0.0",
    "bleach>=6.0.0",
    "python-dateutil>=2.8.2",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.8.0",
    "mypy>=1.0.0",
    "types-PyYAML>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["holespawn", "dashboard"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-v"
filterwarnings = ["ignore::DeprecationWarning"]

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["holespawn", "dashboard", "tests", "scripts"]
exclude = [
    ".venv",
    "venv",
    "__pycache__",
    "outputs",
    "data",
]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "E402",   # module level import not at top (many intentional late imports)
    "E731",   # lambda assignment (style; can refactor to def later)
    "E741",   # ambiguous variable name 'l' (legacy templates)
    "B008",   # do not perform function calls in argument defaults
    "B904",   # raise without from inside except
    "F841",   # unused variable (clean up incrementally)
    "C401",   # set comprehension (style)
    "C416",   # dict comprehension (style)
]

[tool.ruff.lint.isort]
known-first-party = ["holespawn", "dashboard"]

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
exclude = [
    "dashboard/",
    "scripts/",
]

[[tool.mypy.overrides]]
module = "tests.*"
ignore_errors = true
